abstract doc list to separate component
